Test with safe delete tables

+++ release/src/main/groovy/mobilegaming-java-dataflow.groovy
@@ -84,12 +84,18 @@ class LeaderBoardRunner {
     ].join(",")
 
     // Remove existing tables if they exist
-    t.run("bq rm -f -t ${dataset}.${userTable}")
-    t.run("bq rm -f -t ${dataset}.${teamTable}")
+    String tables = t.run("bq query --use_legacy_sql=false 'SELECT table_name FROM ${dataset}.INFORMATION_SCHEMA.TABLES'")
+
+    if (tables.contains(userTable)) {
+      t.run("bq rm -f -t ${dataset}.${userTable}")
+    }
+    if (tables.contains(teamTable)) {
+      t.run("bq rm -f -t ${dataset}.${teamTable}")
+    }
 
     // It will take couple seconds to clean up tables.
     // This loop makes sure tables are completely deleted before running the pipeline
-    String tables = t.run("bq query --use_legacy_sql=false 'SELECT table_name FROM ${dataset}.INFORMATION_SCHEMA.TABLES'")
+    tables = t.run("bq query --use_legacy_sql=false 'SELECT table_name FROM ${dataset}.INFORMATION_SCHEMA.TABLES'")
     while (tables.contains(userTable) || tables.contains(teamTable)) {
       sleep(3000)
       tables = t.run("bq query --use_legacy_sql=false 'SELECT table_name FROM ${dataset}.INFORMATION_SCHEMA.TABLES'")
