+++ release/src/main/groovy/mobilegaming-java-direct.groovy
@@ -79,12 +79,18 @@ def teamSchema = [
         "timing:STRING"
 ].join(",")
 
-t.run("bq rm -f -t ${dataset}.${userTable}")
-t.run("bq rm -f -t ${dataset}.${teamTable}")
+String tables = t.run("bq query --use_legacy_sql=false 'SELECT table_name FROM ${dataset}.INFORMATION_SCHEMA.TABLES'")
+
+if (tables.contains(userTable)) {
+  t.run("bq rm -f -t ${dataset}.${userTable}")
+}
+if (tables.contains(teamTable)) {
+  t.run("bq rm -f -t ${dataset}.${teamTable}")
+}
 
 // It will take a couple of seconds to clean up tables.
 // This loop makes sure tables are completely deleted before running the pipeline
-String tables = t.run("bq query --use_legacy_sql=false 'SELECT table_name FROM ${dataset}.INFORMATION_SCHEMA.TABLES'")
+tables = t.run("bq query --use_legacy_sql=false 'SELECT table_name FROM ${dataset}.INFORMATION_SCHEMA.TABLES'")
 while (tables.contains(userTable) || tables.contains(teamTable)) {
   sleep(3000)
   tables = t.run("bq query --use_legacy_sql=false 'SELECT table_name FROM ${dataset}.INFORMATION_SCHEMA.TABLES'")
@@ -137,7 +143,7 @@ InjectorThread.stop()
 LeaderBoardThread.stop()
 
 if(!isSuccess){
-  t.error("FAILED: Failed running LeaderBoard on DirectRunner")
+    t.error("FAILED: Failed running LeaderBoard on DirectRunner")
 }
 t.success("LeaderBoard successfully run on DirectRunner.")
